pyproject.toml
[project]
name = "visual-chunking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pymupdf>=1.26.4",
    "unstructured[local-inference,pdf]>=0.18.15",
    "reportlab>=4.0.0",
    "argparse>=1.4.0",
    "ruff>=0.14.0",
    "langchain-core>=0.3.78",
    "langchain-unstructured[local]>=0.1.5",
    "python-dotenv>=1.1.1",
    "llama-cloud-services>=0.6.73",
    "docling>=2.61.1",
    "mineru[core]>=2.6.4",
    "chunking-evaluation",
    "notebook>=7.5.0",
    "sentence-transformers>=5.2.0",
    "jupyter>=1.1.1",
    "pyyaml>=6.0.3",
]

[tool.uv.sources]
chunking-evaluation = { git = "https://github.com/matteomrz/chunking_evaluation_fix.git" }
